pyproject.toml
[project]
name = "upscaler"
version = "0.1.0"
description = "SD-based image upscaler — 50-hour capability study vs SUPIR/HYPIR"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "Brad Hinkel" }]
dependencies = [
    "torch>=2.7",
    "torchvision>=0.18",
    "diffusers>=0.27",
    "transformers>=4.40",
    "accelerate>=0.30",
    "safetensors>=0.4",
    "huggingface-hub>=0.24",
    "pillow>=10",
    "numpy>=1.26",
    "tqdm>=4.66",
    "spandrel>=0.4",
    "matplotlib>=3.8",
    "lpips>=0.1.4",
    "piq>=0.8",
    "pandas>=2.2",
    "boto3>=1.34",
    "imagehash>=4.3",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "ruff>=0.6",
    "pre-commit>=3",
    "ipykernel>=6",
    "jupyter>=1",
    "nbclient>=0.10",
    "nbformat>=5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/upscaler"]

# PyTorch + torchvision from the CUDA 12.8 wheel index so the install works on
# Blackwell (sm_120). Without this, the default index ships CPU or older CUDA
# builds and torchvision's C++ ops fail to register against the cu128 torch.
[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B018"]
# Notebook titles/plots use unicode math symbols (×, →, ², ±). They're intentional.
"notebooks/*" = ["RUF001", "RUF002", "RUF003"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
